feat: Implenet NPM modules first draft
