+++ crates/base/src/utils/graph_resolver.rs
@@ -1,12 +1,23 @@
+use std::path::PathBuf;
 use anyhow::{anyhow, bail};
 use deno_core::error::AnyError;
 use deno_core::ModuleSpecifier;
 use eszip::deno_graph;
-use eszip::deno_graph::source::{Resolver, DEFAULT_JSX_IMPORT_SOURCE_MODULE};
+use eszip::deno_graph::source::{Resolver, DEFAULT_JSX_IMPORT_SOURCE_MODULE, NpmResolver, UnknownBuiltInNodeModuleError};
 use import_map::ImportMap;
 use module_fetcher::args::package_json::{PackageJsonDeps, PackageJsonDepsProvider};
 use module_fetcher::util::sync::AtomicFlag;
 use std::sync::Arc;
+use eszip::deno_graph::NpmPackageReqResolution;
+use deno_core::futures::future;
+use deno_core::futures::future::LocalBoxFuture;
+use deno_core::futures::FutureExt;
+use module_fetcher::args::JsxImportSourceConfig;
+use sb_node::is_builtin_node_module;
+use sb_npm::{CliNpmRegistryApi, NpmResolution, PackageJsonDepsInstaller};
+use deno_semver::package::PackageReq;
+use deno_npm::registry::NpmRegistryApi;
+use crate::js_worker::emitter::EmitterFactory;
 
 /// Result of checking if a specifier is mapped via
 /// an import map or package.json.
@@ -104,23 +115,90 @@ pub struct CliGraphResolver {
     maybe_default_jsx_import_source: Option<String>,
     maybe_jsx_import_source_module: Option<String>,
     maybe_vendor_specifier: Option<ModuleSpecifier>,
-    // no_npm: bool,
-    // npm_registry_api: Arc<CliNpmRegistryApi>,
-    // npm_resolution: Arc<NpmResolution>,
-    // package_json_deps_installer: Arc<PackageJsonDepsInstaller>,
+    no_npm: bool,
+    npm_registry_api: Arc<CliNpmRegistryApi>,
+    npm_resolution: Arc<NpmResolution>,
+    package_json_deps_installer: Arc<PackageJsonDepsInstaller>,
     found_package_json_dep_flag: Arc<AtomicFlag>,
+}
+
+pub struct CliGraphResolverOptions<'a> {
+    pub maybe_jsx_import_source_config: Option<JsxImportSourceConfig>,
+    pub maybe_import_map: Option<Arc<ImportMap>>,
+    pub maybe_vendor_dir: Option<&'a PathBuf>,
+    pub no_npm: bool,
+}
+
+impl CliGraphResolver {
+    pub fn new(
+        npm_registry_api: Arc<CliNpmRegistryApi>,
+        npm_resolution: Arc<NpmResolution>,
+        package_json_deps_provider: Arc<PackageJsonDepsProvider>,
+        package_json_deps_installer: Arc<PackageJsonDepsInstaller>,
+        options: CliGraphResolverOptions,
+    ) -> Self {
+        Self {
+            mapped_specifier_resolver: MappedSpecifierResolver {
+                maybe_import_map: options.maybe_import_map,
+                package_json_deps_provider,
+            },
+            maybe_default_jsx_import_source: options
+                .maybe_jsx_import_source_config
+                .as_ref()
+                .and_then(|c| c.default_specifier.clone()),
+            maybe_jsx_import_source_module: options
+                .maybe_jsx_import_source_config
+                .map(|c| c.module),
+            maybe_vendor_specifier: options
+                .maybe_vendor_dir
+                .and_then(|v| ModuleSpecifier::from_directory_path(v).ok()),
+            no_npm: options.no_npm,
+            npm_registry_api,
+            npm_resolution,
+            package_json_deps_installer,
+            found_package_json_dep_flag: Default::default(),
+        }
+    }
+
+    pub fn as_graph_resolver(&self) -> &dyn Resolver {
+        self
+    }
+
+    pub fn as_graph_npm_resolver(&self) -> &dyn NpmResolver {
+        self
+    }
+
+    pub async fn force_top_level_package_json_install(
+        &self,
+    ) -> Result<(), AnyError> {
+        self
+            .package_json_deps_installer
+            .ensure_top_level_install()
+            .await
+    }
+
+    pub async fn top_level_package_json_install_if_necessary(
+        &self,
+    ) -> Result<(), AnyError> {
+        if self.found_package_json_dep_flag.is_raised() {
+            self.force_top_level_package_json_install().await?;
+        }
+        Ok(())
+    }
 }
 
 impl Default for CliGraphResolver {
     fn default() -> Self {
         // This is not ideal, but necessary for the LSP. In the future, we should
         // refactor the LSP and force this to be initialized.
-        // let npm_registry_api = Arc::new(CliNpmRegistryApi::new_uninitialized());
-        // let npm_resolution = Arc::new(NpmResolution::from_serialized(
-        //     npm_registry_api.clone(),
-        //     None,
-        //     None,
-        // ));
+        let emitter_factory = EmitterFactory::new();
+        let npm_registry_api = emitter_factory.npm_api();
+        let npm_resolution = Arc::new(NpmResolution::from_serialized(
+            npm_registry_api.clone(),
+            None,
+            None,
+        ));
+
         Self {
             mapped_specifier_resolver: MappedSpecifierResolver {
                 maybe_import_map: Default::default(),
@@ -129,10 +207,10 @@ impl Default for CliGraphResolver {
             maybe_default_jsx_import_source: None,
             maybe_jsx_import_source_module: None,
             maybe_vendor_specifier: None,
-            // no_npm: false,
-            // npm_registry_api,
-            // npm_resolution,
-            // package_json_deps_installer: Default::default(),
+            no_npm: false,
+            npm_registry_api,
+            npm_resolution,
+            package_json_deps_installer: Default::default(),
             found_package_json_dep_flag: Default::default(),
         }
     }
@@ -182,5 +260,67 @@ impl Resolver for CliGraphResolver {
         }
 
         result
+    }
+}
+
+impl NpmResolver for CliGraphResolver {
+    fn resolve_builtin_node_module(
+        &self,
+        specifier: &ModuleSpecifier,
+    ) -> Result<Option<String>, UnknownBuiltInNodeModuleError> {
+        if specifier.scheme() != "node" {
+            return Ok(None);
+        }
+
+        let module_name = specifier.path().to_string();
+        if is_builtin_node_module(&module_name) {
+            Ok(Some(module_name))
+        } else {
+            Err(UnknownBuiltInNodeModuleError { module_name })
+        }
+    }
+
+    fn load_and_cache_npm_package_info(
+        &self,
+        package_name: &str,
+    ) -> LocalBoxFuture<'static, Result<(), AnyError>> {
+        if self.no_npm {
+            // return it succeeded and error at the import site below
+            return Box::pin(deno_core::futures::future::ready(Ok(())));
+        }
+        // this will internally cache the package information
+        let package_name = package_name.to_string();
+        let api = self.npm_registry_api.clone();
+
+        async move {
+            api
+                .package_info(&package_name)
+                .await
+                .map(|_| ())
+                .map_err(|err| err.into())
+        }.boxed()
+    }
+
+    fn resolve_npm(&self, package_req: &PackageReq) -> NpmPackageReqResolution {
+        if self.no_npm {
+            return NpmPackageReqResolution::Err(anyhow!(
+        "npm specifiers were requested; but --no-npm is specified"
+      ));
+        }
+
+        let result = self
+            .npm_resolution
+            .resolve_package_req_as_pending(package_req);
+        match result {
+            Ok(nv) => NpmPackageReqResolution::Ok(nv),
+            Err(err) => {
+                if self.npm_registry_api.mark_force_reload() {
+                    println!("Restarting npm specifier resolution to check for new registry information. Error: {:#}", err);
+                    NpmPackageReqResolution::ReloadRegistryInfo(err.into())
+                } else {
+                    NpmPackageReqResolution::Err(err.into())
+                }
+            }
+        }
     }
 }
