+++ crates/base/test_cases/eszip-silly-test/index.ts
@@ -1,3 +1,4 @@
 import { is_even } from "https://deno.land/x/is_even@v1.0/mod.ts"
+import isEven from "npm:is-even";
 
-globalThis.isTenEven = is_even(10);
+globalThis.isTenEven = isEven(10);
