+++ crates/module_fetcher/src/lib.rs
@@ -5,7 +5,6 @@ pub mod emit;
 pub mod file_fetcher;
 pub mod http_util;
 pub mod node;
-pub mod npm;
 pub mod permissions;
 pub mod util;
 pub mod version;
