+++ crates/npm/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "sb_npm"
+version = "0.1.0"
+authors = ["Supabase <team@supabase.com>"]
+edition = "2021"
+resolver = "2"
+description = "We'll take care of this later"
+license = "MIT"
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+deno_ast.workspace = true
+deno_core.workspace = true
+deno_npm.workspace = true
+deno_fs.workspace = true
+deno_semver.workspace = true
+once_cell.workspace = true
+async-trait = "0.1.73"
+base32 = "=0.4.0"
+module_fetcher = { path = "../module_fetcher" }
+deno_lockfile.workspace = true
+tar.workspace = true
+flate2.workspace = true
+ring.workspace = true
+sb_node = { version = "0.1.0", path = "../node" }
+serde.workspace = true
+percent-encoding = "=2.3.0"
+hex = "0.4"
+base64 = { version = "=0.13.1" }
+bincode = "=1.3.3"
