+++ crates/sb_core/Cargo.toml
@@ -11,6 +11,7 @@ license = "MIT"
 path = "lib.rs"
 
 [dependencies]
+deno_ast.workspace = true
 deno_net.workspace = true
 deno_web.workspace = true
 deno_fetch.workspace = true
