+++ examples/main-empty/some-import.ts
@@ -1,0 +1,1 @@
+export const sum = (n1, n2) => n1 + n2;
